Moved pixbuf->surface to render module
